refactor(app-shell): split 732-LOC monolith into 8 focused files
Extracts AppShell's responsibilities into app-shell/ submodules. Behavior
preserved 1:1 — no UI, routing, or permission changes. Structure follows
the same convention as rh/ and tasks/ (folder contains everything).

Layout:
  components/app-shell/
  ├── app-shell.tsx               (orchestrator, 120 LOC)
  ├── nav-config.ts               (NAV_ITEMS + path helpers, 136 LOC)
  ├── use-shell-state.ts          (user/clock/cost/collapsed hooks, 135 LOC)
  ├── sidebar.tsx                 (aside + NavSkeleton, 323 LOC)
  ├── header.tsx                  (section title + pills, 121 LOC)
  ├── account-menu.tsx            (avatar + sign-out, 103 LOC)
  ├── info-pill.tsx               (label+value capsule, 11 LOC)
  └── impersonation-banner.tsx    (amber banner, 25 LOC)

Notes:
- AccountMenu now owns its own menuOpen state; it closes on route change
  via its own usePathname effect (matches pre-refactor behavior).
- Sidebar closes the mobile overlay via the parent effect in app-shell.
- The SSR-hydration pattern in use-shell-state.ts (mount-effect reading
  localStorage + setNow) is flagged by react-hooks/set-state-in-effect on
  custom hooks — disabled inline since the pattern is intentional.

Single import site updated: app/layout.tsx.

Followups (not in this PR):
- health-dashboard-view.tsx (757 LOC)
- travel-expense-tracker.tsx (456 LOC)
- components/ui.tsx → components/ui/*
- Tests for /api/impersonate
