+++ components/app-shell/nav-config.ts
@@ -1,0 +1,136 @@
+/**
+ * Navigation data and path helpers for the AppShell.
+ *
+ * Pure module — no React, no side effects. Kept separate so tests, storybooks,
+ * or other shells can reuse the same nav topology without importing the UI tree.
+ */
+
+export type NavChild = {
+  label: string;
+  href: string;
+  divider?: boolean;
+};
+
+export type NavItem = {
+  href: string;
+  labelKey: string;
+  icon: string;
+  matchPaths?: string[];
+  children?: NavChild[];
+};
+
+export const NAV_ITEMS: NavItem[] = [
+  { href: '/', labelKey: 'nav.overview', icon: '🏠' },
+  {
+    href: '/dilesa',
+    labelKey: 'DILESA',
+    icon: 'DILESA_LOGO',
+    children: [
+      { label: 'Administración', href: '#', divider: true },
+      { label: 'Tareas', href: '/dilesa/admin/tasks' },
+      { label: 'Juntas', href: '/dilesa/admin/juntas' },
+      { label: 'Documentos', href: '/dilesa/admin/documentos' },
+      { label: 'Recursos Humanos', href: '#', divider: true },
+      { label: 'Empleados', href: '/dilesa/rh/empleados' },
+      { label: 'Puestos', href: '/dilesa/rh/puestos' },
+      { label: 'Departamentos', href: '/dilesa/rh/departamentos' },
+    ],
+  },
+  {
+    href: '/rdb',
+    labelKey: 'Rincón del Bosque',
+    icon: 'RDB_LOGO',
+    children: [
+      { label: 'Administración', href: '#', divider: true },
+      { label: 'Tareas', href: '/rdb/admin/tasks' },
+      { label: 'Juntas', href: '/rdb/admin/juntas' },
+      { label: 'Documentos', href: '/rdb/admin/documentos' },
+      { label: 'Recursos Humanos', href: '#', divider: true },
+      { label: 'Empleados', href: '/rdb/rh/empleados' },
+      { label: 'Puestos', href: '/rdb/rh/puestos' },
+      { label: 'Departamentos', href: '/rdb/rh/departamentos' },
+      { label: 'Operaciones', href: '#', divider: true },
+      { label: 'Ventas', href: '/rdb/ventas' },
+      { label: 'Cortes', href: '/rdb/cortes' },
+      { label: 'Productos', href: '/rdb/productos' },
+      { label: 'Inventario', href: '/rdb/inventario' },
+      { label: 'Proveedores', href: '/rdb/proveedores' },
+      { label: 'Requisiciones', href: '/rdb/requisiciones' },
+      { label: 'Órdenes de Compra', href: '/rdb/ordenes-compra' },
+      { label: 'Playtomic', href: '/rdb/playtomic' },
+    ],
+  },
+  {
+    href: '/coda',
+    labelKey: 'nav.coda',
+    icon: '📊',
+    children: [
+      { label: 'DILESA', href: '/coda/dilesa' },
+      { label: 'ANSA', href: '/coda/ansa' },
+      { label: 'ANSA-Ventas', href: '/coda/ansa-ventas' },
+      { label: 'SR Group', href: '/coda/sr-group' },
+      { label: 'RDB', href: '/coda/rdb' },
+    ],
+  },
+  {
+    href: '/family',
+    labelKey: 'Familia / Grupo SR',
+    icon: 'SR_LOGO',
+    matchPaths: ['/family', '/travel', '/health'],
+    children: [
+      { label: 'Viajes', href: '/travel' },
+      { label: 'Salud', href: '/health' },
+      { label: 'Familia', href: '/family' },
+    ],
+  },
+  {
+    href: '/settings',
+    labelKey: 'nav.settings',
+    icon: '⚙️',
+    children: [
+      { label: 'Acceso', href: '/settings/acceso' },
+      { label: 'Empresas', href: '/settings/empresas' },
+      { label: 'Integraciones', href: '/settings/integraciones' },
+      { label: 'Preferencias', href: '/settings/preferencias' },
+    ],
+  },
+];
+
+/** Maps top-level nav hrefs to their empresa slug for permission filtering. */
+export const NAV_TO_EMPRESA: Record<string, string> = {
+  '/dilesa': 'dilesa',
+  '/rdb': 'rdb',
+  '/coda': 'coda',
+  '/family': 'familia',
+  '/settings': 'settings',
+};
+
+export function matchesPath(pathname: string, href: string) {
+  return href === '/' ? pathname === '/' : pathname === href || pathname.startsWith(`${href}/`);
+}
+
+export function isItemActive(pathname: string, item: NavItem) {
+  const paths = item.matchPaths ?? [item.href];
+  return paths.some((path) => matchesPath(pathname, path));
+}
+
+export function getActiveSection(pathname: string) {
+  return NAV_ITEMS.find((item) => item.children && isItemActive(pathname, item))?.href ?? null;
+}
+
+export function getSectionLabelKey(pathname: string) {
+  return NAV_ITEMS.find((item) => isItemActive(pathname, item))?.labelKey ?? 'nav.overview';
+}
+
+export function getInitials(name: string, email: string) {
+  const source = name.trim() || email.trim();
+  const parts = source.split(/\s+/).filter(Boolean).slice(0, 2);
+  if (parts.length === 0) return 'BS';
+  return parts.map((part) => part[0]?.toUpperCase() ?? '').join('');
+}
+
+export type AuthUser = {
+  name: string;
+  email: string;
+  avatarUrl: string | null;
+};
